+++ nczarr_test/ut_util.h
@@ -6,6 +6,7 @@
 #ifndef UT_UTIL_H
 #define UT_UTIL_H
 
+#include <stddef.h>
 extern int parseslices(const char* s0, int* nslicesp, NCZSlice* slices);
 extern int parsedimdef(const char* s0, Dimdef** defp);
 extern int parsevardef(const char* s0, NClist* dimdefs, Vardef** varp);
@@ -17,7 +18,7 @@ extern void freeranges(NCZChunkRange* ranges);
 extern void freeslices(NCZSlice* slices);
 extern void freestringvec(char** vec);
 extern void freeprojvector(int rank, NCZProjection** vec);
-extern int ut_typesize(nc_type t);
+extern size_t ut_typesize(nc_type t);
 extern nc_type ut_typeforname(const char* tname);
 extern NCZM_IMPL kind2impl(const char* kind);
 extern const char* impl2kind(NCZM_IMPL impl);
